% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTRNG-stickR-app.R
\name{rTRhexNG_app}
\alias{rTRhexNG_app}
\title{rTRhexNG Shiny App}
\usage{
rTRhexNG_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
rTRhexNG Shiny App
}
